Revert "Add the "Tech News" sub-category"
This reverts commit 20f0a3a41652f7e5f45b639be8d250236a9ab004.

"Tech News" was never an official podcast category as far as Apple was concerned. They recently announced new podcast categories rolling out sometime soon. In them, "Tech News" is real, but it's a sub-category to "News". While the Haskell Weekly Podcast tries to be timely, it's not really news. I think that "Technology" is a better overall description of the content.

https://podnews.net/article/apple-changed-podcast-categories-2019